+++ src/app/Home/_layout.tsx
@@ -1,0 +1,15 @@
+import { Stack } from 'expo-router';
+
+export default function HomeLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+        animation: 'slide_from_right',
+        contentStyle: {
+          backgroundColor: '#0F2537',
+        },
+      }}
+    />
+  );
+}